Layout + vorig donderdag (#25)
* Hotfix 

Deletion of } accidentally

* header multipurpose

Changing header to multipurpose for patient gallery and also for scans gallery. In patient gallery in the header is showed: "Welcome Dr...". In the Scan Gallery is showed: "Scans of patient Firstname Lastname under supervision of Dr .. "

* patient redirection to scan gallery page completed

In the patient Gallery page, when clicking on the Image of a certain patient. It sends to the scans gallery page. in the URL is the first name and last name of the patient passed.